Crate inth_oauth2

"It's not that hard" OAuth 2.0

OAuth 2.0 really isn't that hard, you know?

Implementation of RFC 6749.

inth_oauth2 is on Crates.io and GitHub.

Providers

Support for the following OAuth 2.0 providers is included:

Google

Web

Installed

GitHub

Imgur

Support for other providers can be added by implementing the Provider trait.

Token types

The only supported token type is Bearer. Support for others can be added by implementing the Token trait.

Examples

Creating a client

use inth_oauth2::Client;
use inth_oauth2::provider::google::Installed;

let client = Client::<Installed>::new(
    String::from("client_id"),
    String::from("client_secret"),
    Some(String::from("redirect_uri"))
);

Constructing an authorization URI

let auth_uri = client.auth_uri(Some("scope"), Some("state")).unwrap();
println!("Authorize the application by clicking on the link: {}", auth_uri);

Requesting an access token

use std::io;
use inth_oauth2::{Client, Token};

let mut code = String::new();
io::stdin().read_line(&mut code).unwrap();

let http_client = Default::default();
let token = client.request_token(&http_client, code.trim()).unwrap();
println!("{}", token.access_token());

Refreshing an access token

let token = client.refresh_token(&http_client, token, None).unwrap();

Ensuring an access token is still valid

// Refresh token only if it has expired.
token = client.ensure_token(&http_client, token).unwrap();

Using bearer access tokens

Bearer tokens can be converted to Hyper headers.

use hyper::header::Authorization;

let client = hyper::Client::new();
let request = client.get("https://example.com/resource")
    .header(Into::<Authorization<_>>::into(&token));

Persisting tokens

All token types implement Encodable / Decodable from rustc_serialize and Serialize / Deserialize from serde (with the default serde feature).

use rustc_serialize::json;
let json = json::encode(&token).unwrap();

extern crate serde_json;
let json = serde_json::to_string(&token).unwrap();

Reexports

pub use token::{Token, Lifetime};

pub use client::{Client};

Modules

client
Client.

error
Errors.

provider
Providers.

token
Tokens.

Enums

ClientError
Errors that can occur during authorization.